casting to int

+++ example_c_atsamg55XplainedPRO_atmelStudio/src/main.c
@@ -110,10 +110,10 @@ int exolib_read_complete(Exosite_state_t *exo, int status)
 
 int exolib_timestamp_complete(Exosite_state_t *exo, uint32_t timestamp)
 {
-    printf(":: Got timestamp %d.\r\n", timestamp);
+    printf(":: Got timestamp %d.\r\n", (int)timestamp);
     // !!! Remember, the buffer needs to exist until the write callback is called.
     // So you cannot use memory on the stack.
-    snprintf(writeReqBuffer, sizeof(writeReqBuffer), "change=%d", timestamp);
+    snprintf(writeReqBuffer, sizeof(writeReqBuffer), "change=%d", (int)timestamp);
     exosite_write(exo, writeReqBuffer);
     return 0;
 }
